parse sky scroll speed as double

+++ Core/Resources/Definitions/MapInfo/SkyDef.cs
@@ -5,7 +5,7 @@ namespace Helion.Resources.Definitions.MapInfo;
 public class SkyDef : ICloneable
 {
     public string Name { get; set; } = string.Empty;
-    public int ScrollSpeed { get; set; }
+    public double ScrollSpeed { get; set; }
 
     public object Clone()
     {
